+++ miniprogram/app.js
@@ -1,0 +1,8 @@
+App({
+  onLaunch() {
+    console.log('Project-X 小程序启动');
+  },
+  globalData: {
+    webUrl: ''
+  }
+});